+++ pkg/config.go
@@ -18,6 +18,7 @@ const (
 	Runner      = "runner"
 	GroupRunner = "groupRunner"
 	User        = "user"
+	Pipeline    = "pipeline"
 
 	// GitLab API OK Response Code
 	gitlabReponseCodeOk = 204
@@ -34,8 +35,9 @@ type Config struct {
 	Insecure              bool          `goptions:"--insecure, description='Skip certificate Verfication for Gitlab API URL, (bool)'"`
 	Token                 string        `goptions:"-t, --token, description='Gitlab API access token, can also be set via env \"GITLAB_TOKEN\" or file \".token\"'"`
 	Operation             string        `goptions:"-o, --operation, description='Operation to start, (list, archive, archive-all, delete, delete-all)', obligatory"`
-	Resource              string        `goptions:"-r, --resource, description='Resource to interact with, (project, runner, groupRunner, user)', obligatory"`
+	Resource              string        `goptions:"-r, --resource, description='Resource to interact with, (project, runner, groupRunner, user, pipeline)', obligatory"`
 	ResourceId            int           `goptions:"-i, --identifier, description='Resource ID to interact with, (int)'"`
+	PipelineId            int           `goptions:"--pipelineId, description='Pipeline ID to interact with, (int)'"`
 	ProjectType           string        `goptions:"-p, --project-type, description='Type of project (internal, private, public), (default: internal), (string)'"`
 	Age                   int           `goptions:"-a, --age, description='Filter by last activity in months (not available for runner), (int)'"`
 	Query                 string        `goptions:"-q, --query, description='Search by name, (string)'"`